wshnationalsbot/requirements.txt
python-telegram-bot[job-queue]==20.3
MLB-StatsAPI==1.9.0
pytz==2024.1
python-dotenv==1.0.1
flask==3.0.3
requests==2.31.0
Pillow>=10.0.0
